Refactored standalone UAP test class using @Nested.
